Add blueprint modules and wrappers
